move parameter.xpp under server/
